Natural Ergonomic Keyboard 4000: make zoom slider work
This keyboard has a zoom slider. Unfortunately it does nothing, when
shifted (in the former configuration). And it is not intuitive that it is
a 'zoom slider'. Most people would expect it to be a 'scroll slider'.
Which is proven by many posts on forums in which people ask how to make
the 'scroll slider' work.
It is now working as expected.