MS Natural Ergonomic Keyboard 4000: Change the zoom slider to act like the up/down keys
This keyboard has a zoom slider. Unfortunately linux software has almost
no support for zooming at desktop level. This leads to the assumption
that this slider is not working. Additional people think intuitively
that the zoom slider is indeed a 'scroll slider'.This can be seen in
many posts in forums in which people ask "how to get the 'scroll slider'
working".
I have changed the settings of this particular keyboard that the slider
acts now like the up and down keys. So e.g. a website that is higher
than the screen can be read by tilting the slider downward.