seccomp: add @basic-rw syscall set
This adds @basic-rw syscall set that includes all read, write and lseek
syscalls. This allowes to support strict sandboxes like seccomp v1 or
setups where apps or services are not allowed to open, dup or close file
destricptors, only work with the received file descriptors.

The group is imported in @basic-io for backward compatiblity of course
but also @basic-io is the super set of this one.