seccomp: add @basic-memory syscall set
This adds @basic-memory mapping syscall sets. With this set we separate
the related syscalls from the @file-system set, however it is imported
into @file-system to not break backward compatibility and more ever
that's were the syscalls belong. The use of @basic-memory allows to
create restricted sandboxes without whitelisting all @file-system
syscalls.