seccomp: add the '@basic-directory' syscall set
Add the @basic-directory syscall set that gets the current working
directory or changes it. This one does not include fchdir() which will
be in a nother set that workds with file descriptors. Separating the set
from @file-system syscall set allows to easily use this set without
whitelisting all file system operations, however it is included in that
set as that set should contain all filesystem operations.