signal-util: add a wrapper for SIGRTMIN
SIGRTMIN is a macro which expands to __libc_current_sigrtmin() which
does just returns current_rtmin, which is a static int variable. It
can be incremented using __libc_allocate_rtsig(true).

If __libc_allocate_rtsig() would be called at any time except very early,
chaos would ensue. In particular we register signal handlers for specific
numeric values, and if those values ever changed, we would call different
actions than expected. The effect is that from our point of view, SIGRTMIN
is always constant, but gcc must treat it as an arbitrary functions and
always call __libc_current_sigrtmin().

To avoid this, let's add a global variable which is initialized early.
The code that gcc generates is changed from (at -O0):

    13f5:       e8 a6 f6 ff ff          callq  aa0 <.plt.got+0x60>
    13fa:       83 c0 05                add    $0x5,%eax
    13fd:       39 45 fc                cmp    %eax,-0x4(%rbp)
    1400:       74 27                   je     1429 <test_sigrtmin_comparisons_orig+0x45>
    1402:       e8 99 f6 ff ff          callq  aa0 <.plt.got+0x60>
    1407:       83 c0 06                add    $0x6,%eax
    140a:       39 45 fc                cmp    %eax,-0x4(%rbp)
    140d:       74 1a                   je     1429 <test_sigrtmin_comparisons_orig+0x45>
    140f:       e8 8c f6 ff ff          callq  aa0 <.plt.got+0x60>
    1414:       83 c0 0f                add    $0xf,%eax
    1417:       39 45 fc                cmp    %eax,-0x4(%rbp)
    141a:       74 0d                   je     1429 <test_sigrtmin_comparisons_orig+0x45>
    141c:       e8 7f f6 ff ff          callq  aa0 <.plt.got+0x60>
    1421:       83 c0 10                add    $0x10,%eax
    1424:       39 45 fc                cmp    %eax,-0x4(%rbp)

to

    1444:       8b 05 c6 0b 20 00       mov    0x200bc6(%rip),%eax        # 202010 <sigrtmin@@SD_SHARED>
    144a:       83 c0 05                add    $0x5,%eax
    144d:       39 45 fc                cmp    %eax,-0x4(%rbp)
    1450:       74 2a                   je     147c <test_sigrtmin_comparisons+0x45>
    1452:       8b 05 b8 0b 20 00       mov    0x200bb8(%rip),%eax        # 202010 <sigrtmin@@SD_SHARED>
    1458:       83 c0 06                add    $0x6,%eax
    145b:       39 45 fc                cmp    %eax,-0x4(%rbp)
    145e:       74 1c                   je     147c <test_sigrtmin_comparisons+0x45>
    1460:       8b 05 aa 0b 20 00       mov    0x200baa(%rip),%eax        # 202010 <sigrtmin@@SD_SHARED>
    1466:       83 c0 0f                add    $0xf,%eax
    1469:       39 45 fc                cmp    %eax,-0x4(%rbp)
    146c:       74 0e                   je     147c <test_sigrtmin_comparisons+0x45>
    146e:       8b 05 9c 0b 20 00       mov    0x200b9c(%rip),%eax        # 202010 <sigrtmin@@SD_SHARED>
    1474:       83 c0 10                add    $0x10,%eax
    1477:       39 45 fc                cmp    %eax,-0x4(%rbp)

Essentialy, callq is replaced by mov with a global offset.

At -O3, the "new" code becomes very hard to follow, because the compiler is
able to coalesce/rearrange the comparisons. The "old" code doesn't change much,
all callq's are still present.

At -O3, the size of 'systemd' binary is reduced by a few hundred bytes. This is
a bit disappointing, but I think the small difference is caused by the fact
that mov is actually encoded in a higher number of bytes (6 vs. 5).
Nevertheless, a simple mov should be faster then a call through plt.