Extend ConfigureWithoutCarrier to carrier changes
Recently new [Network] option ConfigureWithoutCarrier was added which configures the link even if it has no carrier.

The idea is that IP addresses / routes are configured even if there is no carrier at the start of systemd-networkd service.

But in current case, if carrier is lost later (even for a moment) then this configuration (IP address / routes) are dropped. Which is obviously not what one wants. The purpose of the patch is that configuration exists regardless of carrier state.

This patch handles such cases. If link is already configured and if ConfigureWithoutCarrier is set then no action is taken on carrier gain or loss. (i.e. configuration remains as is)