core: in setup_keyring() make chown operations non-fatal
If kernel denies changing the ownership of the keyring, simply remove the
invocation_id key, and do not error out of the function. This makes
setup_keyring() non-fatal in such a case.

Changing ownership of the keyring may fail, if target uid exceeds quota,
KEYCTL_CHOWN is filtered with seccomp/LSM, etc.

Fixes #6281