Use 'sulogin --force' instead of 'sulogin'
When the root account is locked, which is the case for almost all
Ubuntu installations and (presumably) many Debian installations,
sulogin will refuse to start a root shell.

sulogin introduced the option '--force' for systems with locked
root accounts. Note that 'sulogin --force' should still prompt for
the root account password when the root account is not locked.