udev: net_id: assign path based name to non pci based usb network devices
USB network devices are getting predictable names beeing prefixed with
the PCI root path, but embedded usb controller are not necessarily
connected using PCI, so these network devices keep getting named ethX.

This patch addresses this issue and introduces a naming scheme for those
non pci based usb network controllers. When no PCI controller is found
the usb device name is build up like normal, but instead of the enpXsYuZ
name the are assigned enUXuZ names, where the X is the number of the
usb roothub, gathered from the same string which also has the hub port
numbers encoded. When the usb-device is connected via PCI the old naming
schema is kept.